requirements.txt
# --- Core Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1  

# --- Data Processing & Science ---
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
Pillow>=9.5.0

# --- Visualization & Progress ---
matplotlib>=3.7.0
tqdm>=4.65.0

# --- Web API & Deployment ---
fastapi
uvicorn
python-multipart  